Created Input.java
In this code I wrote a line outside of the class that imports a java utility called "Scanner". This is a method for getting input from the user and the input is then echoed back to the user. At this point in time I am not sure what the terms that come before "main" mean or why they matter and I have a lot of questions. This program seems fairly useless and the first thing I would want to change about is it make it so I can repeatedly type something in and have it echoed back rather than having to completely restart the program each time.