Code for Basic Calculator
This is a program that can only add two numbers and then the program is terminated after doing so. I once again imported the java.util.Scanner in order to be able to input information from my keyboard into the terminal.

On line 5 my understanding of this piece of code is that "Scanner x" means that there is a variable that has a Scanner data type. That variable is updated by whatever new Scanner input is entered on System.in (the terminal?).

Line 6 is basic variables as doubles which means they can have decimal points.

Line 8 I don'y fully understand what is meant by "nextDouble" unless that means that like once that value that x holds is assigned to fnum it is then dropped and ready for a new value on to be entered which snum then takes?