Using Methods with Parameters
In this code converts another class into an object and then class for a method from that object. at the start of the main function on line 5 there is a Scanner variable created called "input" and it is set to equal the input from the user.
Line 8-11 is a basic interface with a String variable named "name" that takes on the string that the user types in and then that variable is used at the parameter for the simpleMessage method from the object.